+++ R/oce2ncdf.R
@@ -1,6 +1,6 @@
 #' Save an oce-class object as a netcdf file.
 #'
-#' `oce1ncdf()` works by determining the class of its first argument,
+#' `oce2ncdf()` works by determining the class of its first argument,
 #' and then dispatching to an internal function, as appropriate.
 #'
 #' @param x an oce-class object. The subclass of this object determines
@@ -34,6 +34,7 @@ oce2ncdf <- function(x, varTable=NULL, ncfile=NULL, debug=0)
     switch(xclass,
         ctd=ctd2ncdf(x, varTable=varTable, ncfile=ncfile, debug=debug),
         adp=adp2ncdf(x, varTable=varTable, ncfile=ncfile, debug=debug),
+        adv=adv2ncdf(x, varTable=varTable, ncfile=ncfile, debug=debug),
         stop("oce2ncdf() cannot handle \"", xclass, "\" objects")
     )
 }